Move ante test suite execution to ante_test
